refactor(profile): separate menu as component
