+++ src/components/layouts/layout-profile/layout-profile-content/layout-profile-content.module.scss
@@ -1,3 +1,12 @@
+@use "../../../../styles/variables";
+
 .content {
+  padding-top: 3rem;
   grid-area: content;
+}
+
+@media (min-width: variables.$breakpoint-md) {
+  .content {
+    padding-top: initial;
+  }
 }
