+++ src/components/layouts/layout-profile/layout-profile-menu/layout-profile-menu.module.scss
@@ -1,0 +1,24 @@
+@use "../../../../styles/variables";
+@use "../../../../styles/colors";
+
+.button {
+  composes: text text_type_main-default text_color_inactive from global;
+  border: none;
+  background: none;
+  cursor: pointer;
+
+  &.active {
+    color: colors.$white;
+  }
+
+  &:hover:not(.active) {
+    text-shadow: 0 0 5px colors.$violet;
+  }
+}
+
+.navList {
+  display: flex;
+  flex-flow: column nowrap;
+  justify-content: space-around;
+  gap: 1em;
+}
