+++ src/components/layouts/layout-profile/layout-profile-nav/layout-profile-nav.module.scss
@@ -2,39 +2,15 @@
 @use "../../../../styles/colors";
 
 .nav {
-  composes: text_type_main-medium text_color_inactive mb-10 from global;
-  grid-area: nav;
-  inline-size: 100%;
-}
-
-.button {
-  composes: text text_type_main-medium text_color_inactive from global;
-  border: none;
-  background: none;
-  cursor: pointer;
-
-  &.active {
-    color: colors.$white;
-  }
-
-  &:hover:not(.active) {
-    text-shadow: 0 0 5px colors.$violet;
-  }
-}
-
-.navList {
-  display: flex;
-  flex-flow: row nowrap;
-  justify-content: space-around;
+  display: none;
 }
 
 @media (min-width: variables.$breakpoint-md) {
   .nav {
+    composes: text_type_main-medium text_color_inactive mb-10 from global;
+    display: initial;
+    grid-area: nav;
+    inline-size: 100%;
     margin-block: variables.$profile-edit-top-space;
-  }
-
-  .navList {
-    flex-flow: column nowrap;
-    gap: 1em;
   }
 }
